+++ packages/camera/camera_platform_interface/lib/src/events/device_event.dart
@@ -60,7 +60,7 @@ class DeviceTiltsChangedEvent extends DeviceEvent {
   final int horizontalTilt;
 
   /// The new vertical tilt of the device
-  final int verticalTilt;
+  final double verticalTilt;
 
   /// Build a new tilts changed event.
   DeviceTiltsChangedEvent({
